Docker-Flask Deployment/app/requirements.txt
Flask==2.3.3
tensorflow==2.13.0
Pillow